Allow custom button mappings

+++ plugin.program.remote.control.browser/retrieve.py
@@ -20,19 +20,21 @@ def main():
     # The image must be re-opened after verification.
     buffered.seek(0)
     image = PIL.Image.open(buffered)
-    with tempfile.NamedTemporaryFile(delete=False) as saved:
-        image.save(saved, PIL.PngImagePlugin.PngImageFile.format)
-
-    # Atomically move to the desired location. (This is only atomic if the
-    # temp directory is on the same filesystem).
+    saved = None
     try:
+        with tempfile.NamedTemporaryFile(delete=False) as saved:
+            image.save(saved, PIL.PngImagePlugin.PngImageFile.format)
+
+        # Atomically move to the desired location. (This is only atomic if the
+        # temp directory is on the same filesystem).
         shutil.move(saved.name, args.filename)
-    except OSError:
-        try:
-            os.remove(saved.name)
-        except OSError:
-            pass
-        raise
+        saved = None
+    finally:
+        if saved is not None:
+            try:
+                os.remove(saved.name)
+            except OSError:
+                pass
 
 
 if __name__ == "__main__":
